% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DoseRate.R
\name{calc_DoseRate}
\alias{calc_DoseRate}
\title{Calculate the dose rate from a spectrum}
\usage{
calc_DoseRate(data, energy.min = get_EnergyThreshold(), energy.max = NULL,
  background.correction = TRUE, plot = TRUE, plot.single = TRUE, ...)
}
\arguments{
\item{data}{input data}

\item{energy.min}{starting energy in eV}

\item{energy.max}{upper integration limit (not implemented)}

\item{background.correction}{subtract background}

\item{plot}{plot}

\item{plot.single}{single plots}

\item{...}{Additional arguments: \code{verbose}}
}
\value{
Returns terminal output, a plot and a \code{\link{list}} of the following 
structure:
}
\description{
Calculate the dose rate from a spectrum
}
\examples{

# none available yet

}
