% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaSpec-package.R
\docType{package}
\name{gammaSpec-package}
\alias{gammaSpec-package}
\title{Analyse Gamma Spectra of NaI(Tl) Scintillation Radiation Detectors}
\description{
A minimal package to import and plot SPE files recorded by
ORTEC NaI(Tl) Scintillation Radiation Detectors with a digiBASE 
Digital Multichannel Analyser as well as to derive the external wet 
gamma dose rate.
}
\author{
Christoph Burow (University of Cologne, Germany)
}
\keyword{package}
