% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaSpec-package.R
\docType{data}
\name{MassAttenuationCoefficients}
\alias{MassAttenuationCoefficients}
\title{Photon Cross Section Data}
\format{A \code{\link{list}} with two elements (\code{\link{data.frame}s})
with two columns (\code{energy_MeV, attenuation_cm2g}: \cr
\tabular{ll}{
$NaI \tab Total attenuation for sodium iodide \cr
$SiO2 \tab Total attenuation for silicon dioxide \cr
}}
\source{
http://physics.nist.gov/PhysRefData/Xcom/html/xcom1-t.html
}
\description{
Total attenuation (with coherent scatter) of photons for NaI and SiO2, given
in cm^2/g. Data taken from the NIST XCOM Photon Cross Section Database.
}
\examples{
data(MassAttenuationCoefficients, envir = environment())
plot(MassAttenuationCoefficients)

}
\keyword{datasets}
