% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MassAttenuationCoefficients.R
\name{plot_MassAttenuationCoefficients}
\alias{plot_MassAttenuationCoefficients}
\title{Plot NaI and SiO2 Mass Attenuation Coefficients}
\usage{
plot_MassAttenuationCoefficients(data, ...)
}
\arguments{
\item{data}{\code{\link{list}} \bold{required}: SiO2 and NaI total attenuation
coefficients included in this package (see examples).}

\item{...}{further parameters passed to \code{\link{plot}}}
}
\value{
Returns a plot.
}
\description{
This function produces a plot of the total mass attenuation coefficients for
NaI divided by the total mass attenuation coefficients for SiO2. Data taken 
from the NIST XCOM Photon Cross Section Database.
}
\details{
<placeholder>
}
\examples{

data("MassAttenuationCoefficients")
plot_MassAttenuationCoefficients(MassAttenuationCoefficients)

}

