% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Spectrum.R
\name{plot_Spectrum}
\alias{plot_Spectrum}
\title{Plot gamma spectrum}
\usage{
plot_Spectrum(data, type = c("bar", "line", "point"), xval = c("energy",
  "channel"), yval = c("rate", "count"), info = TRUE, fill = TRUE, ...)
}
\arguments{
\item{data}{\code{\link{list}} \bold{required}: data object produced by 
\code{\link{read_SPE}}.}

\item{type}{\code{\link{character}} (with default):}

\item{xval}{\code{\link{character}} (with default):}

\item{yval}{\code{\link{character}} (with default):}

\item{info}{\code{\link{character}} (with default):}

\item{fill}{\code{\link{character}} (with default):}

\item{...}{currently not used.}
}
\value{

}
\description{
<placeholder>
}
\details{
<placeholder>
}
\examples{

# none available yet

}

