% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_DoseRate.R
\name{calc_DoseRate}
\alias{calc_DoseRate}
\title{Calculate the dose rate from a spectrum}
\usage{
calc_DoseRate(data, energy.min = get_EnergyThreshold(), ...)
}
\arguments{
\item{data}{}

\item{energy.min}{}

\item{...}{}
}
\value{

}
\description{
Calculate the dose rate from a spectrum
}
\examples{

# none available yet

}

