% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaSpec-package.R
\docType{package}
\name{gammaSpec-package}
\alias{gammaSpec-package}
\title{A Collection of Functions to Analyse Gamma Spectrums}
\description{
<placeholder>
}
\author{
Christoph Burow (University of Cologne, Germany)
}
\keyword{package}

