% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_SPE.R
\name{read_SPE}
\alias{read_SPE}
\title{Import SPE files to R}
\usage{
read_SPE(file, ...)
}
\arguments{
\item{file}{\code{\link{character}} \bold{required}: file path}

\item{...}{currently not used.}
}
\value{

}
\description{
<placeholder>
}
\details{
<placeholder>
}
\examples{

# none available yet

}

